style: redesign UI with minimal layout
- restyle navigation, product cards and product pages
- restyle cart and admin panel
- add shared UI components

+++ online-shop/lib/validations.ts
@@ -3,7 +3,9 @@ import { z } from "zod";
 const name = z.string().min(3).max(100);
 const slug = z.string().regex(/^[a-z-0-9-]+$/).min(3).max(100);
 const description = z.string().min(3).max(500).optional();
-const price = z.coerce.number().int().positive();
+// Coloana din DB e `doublePrecision`, iar prețurile reale au zecimale (5499.99),
+// deci validarea nu trebuie să ceară întregi.
+const price = z.coerce.number().positive();
 const stock = z.coerce.number().int().nonnegative();
 const categoryId = z.coerce.number().int().positive();
